+++ projectroles/templatetags/projectroles_tags.py
@@ -109,8 +109,8 @@ def allow_project_creation():
 
 
 @register.simple_tag
-def is_app_link_visible(plugin, project, user):
-    """Check if app link should be visible for user in a specific project"""
+def is_app_visible(plugin, project, user):
+    """Check if app should be visible for user in a specific project"""
     can_view_app = user.has_perm(plugin.app_permission, project)
     app_hidden = False
 
@@ -137,6 +137,7 @@ def is_app_link_visible(plugin, project, user):
 @register.simple_tag
 def get_project_list(user, parent=None):
     """Return flat project list for displaying in templates"""
+    # TODO: Remove once reimplementing custom column retrieval
     project_list = []
 
     if user.is_superuser:
@@ -223,49 +224,9 @@ def get_not_found_alert(project_results, app_search_data, search_type):
 
 
 @register.simple_tag
-def get_project_list_columns():
-    """Return custom project list columns as a sorted list"""
-    cols = []
-    i = 0
-
-    for app_plugin in [
-        ap
-        for ap in get_active_plugins(plugin_type='project_app')
-        if ap.project_list_columns
-    ]:
-        for k, v in app_plugin.project_list_columns.items():
-            v['app_plugin'] = app_plugin
-            v['key'] = k
-            v['ordering'] = v.get('ordering') or i
-            cols.append(v)
-            i += 1
-
-    return sorted(cols, key=lambda x: x['ordering'])
-
-
-@register.simple_tag
-def get_project_list_value(app_plugin, column_id, project, user):
-    ret = app_plugin.get_project_list_value(column_id, project, user)
-    return ret if ret is not None else ''
-
-
-@register.simple_tag
-def get_project_column_count(app_plugins):
-    """Return the amount of columns shown in project listings"""
-
-    def get_active_list_columns(app_plugin):
-        return len(
-            [
-                column
-                for column, attributes in app_plugin.project_list_columns.items()
-                if attributes['active']
-            ]
-        )
-
-    return 2 + max(
-        [get_active_list_columns(app_plugin) for app_plugin in app_plugins],
-        default=0,
-    )
+def get_project_column_value(col, project):
+    """Return value of a custom project list column for a project"""
+    return col['data'].get(str(project.sodar_uuid))
 
 
 # TODO: Update tests